Rename LogDTO to DpsReportDTO
