+++ src/main/java/qtlog/DataModel/IDataModel.java
@@ -1,10 +1,10 @@
 package qtlog.DataModel;
 
+import qtlog.DPSReportController.DpsReportDTO;
 import qtlog.UIController.IModelObserver;
-import qtlog.shared.LogDTO;
 
 public interface IDataModel {
-    public LogDTO getLatestLog();
+    public DpsReportDTO getLatestLog();
 
     public void registerObs(IModelObserver observer);
 }
