+++ src/main/java/qtlog/UserInterface/UserInterface.java
@@ -1,11 +1,11 @@
 package qtlog.UserInterface;
 
-import qtlog.shared.LogDTO;
+import qtlog.DPSReportController.DpsReportDTO;
 
 public class UserInterface implements IUserInterface {
 
     @Override
-    public void updateUI(LogDTO logInfo) {
+    public void updateUI(DpsReportDTO logInfo) {
         System.out.println("UserInterface updated!\nBoss: " + logInfo.getBossName());
     }
 }
